Consolidated hash drivers.
